+++ packages/core/src/components/avatar/index.ts
@@ -1,0 +1,1 @@
+export { default as VfAvatar } from './VfAvatar.vue';