+++ packages/core/src/entries/avatar.ts
@@ -1,0 +1,3 @@
+import '../styles/entries/avatar.css';
+
+export { VfAvatar } from '../components/avatar';